Refactor function custom backfill
